Tools/Configurator/requirements.txt
PySide6>=6.5
pyserial>=3.5
pyqtgraph>=0.13
numpy>=1.24
